requirements.txt
langchain
openai
google-search-results
googlemaps
python-dotenv